Initial Commit Korrektur 2

+++ Projekt_XancakeLib/demo/de/xancake/persistence/bind/produkt/ProduktBinding.java
@@ -1,5 +1,6 @@
 package de.xancake.persistence.bind.produkt;
 
+import java.math.BigDecimal;
 import de.xancake.persistence.bind.AttributeBinding;
 import de.xancake.persistence.bind.Binding;
 
@@ -9,7 +10,7 @@ public class ProduktBinding extends Binding<Produkt> {
 	private static final AttributeBinding<Integer> ATT_ID         = new AttributeBinding<>("id", Integer.class, false);
 	private static final AttributeBinding<String> ATT_NUMMER      = new AttributeBinding<>("nummer", String.class, false);
 	private static final AttributeBinding<String> ATT_BEZEICHNUNG = new AttributeBinding<>("bezeichnung", String.class, false);
-	private static final AttributeBinding<Double> ATT_PREIS       = new AttributeBinding<>("preis", Double.class, false);
+	private static final AttributeBinding<BigDecimal> ATT_PREIS   = new AttributeBinding<>("preis", BigDecimal.class, false);
 	
 	public ProduktBinding() {
 	    super(ENTITY, ATT_ID, new Factory<Produkt>() {
